+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.2.0
+
+* Add `contentUri` based VideoPlayerController.
+
 ## 2.1.15
 
 * Ensured seekTo isn't called before video player is initialized. Fixes [#89259](https://github.com/flutter/flutter/issues/89259).